Authorization

+++ server/api/products.js
@@ -1,5 +1,6 @@
+/* eslint-disable complexity */
 const router = require('express').Router()
-const {Product} = require('../db/models')
+const {Product, User} = require('../db/models')
 module.exports = router
 
 router.get('/', async (req, res, next) => {
@@ -14,6 +15,7 @@ router.get('/', async (req, res, next) => {
 
 router.get('/:productId', async (req, res, next) => {
   try {
+    console.log(req.session)
     const productById = await Product.findByPk(req.params.productId)
     res.json(productById)
     if (!productById) res.sendStatus(404)
@@ -22,17 +24,29 @@ router.get('/:productId', async (req, res, next) => {
   }
 })
 
-// still need to permission this but need to refer to auther workshop solution
+// eslint-disable-next-line complexity
 router.put('/:productId', async (req, res, next) => {
   try {
-    console.log(req)
-    let productById = await Product.findByPk(req.params.productId)
-    if (!productById) {
-      res.sendStatus(404)
+    if (req.session.passport) {
+      const user = await User.findByPk(req.session.passport.user)
+      const isAnAdmin = user.isAdmin
+      if (isAnAdmin) {
+        let productById = await Product.findByPk(req.params.productId)
+        if (!productById) {
+          res.sendStatus(404)
+        } else {
+          const formData = {}
+          formData.id = req.body.id
+          if (req.body.name) formData.name = req.body.name
+          if (req.body.price) formData.price = req.body.price
+          if (req.body.description) formData.description = req.body.description
+          if (req.body.image) formData.image = req.body.image
+          const updatedProduct = await productById.update(formData)
+          res.send(updatedProduct)
+        }
+      }
     } else {
-      console.log('req.body', req.body)
-      const updatedProduct = await productById.update(req.body)
-      res.send(updatedProduct)
+      res.sendStatus(403)
     }
   } catch (error) {
     next(error)
@@ -41,15 +55,28 @@ router.put('/:productId', async (req, res, next) => {
 
 router.delete('/:productId', async (req, res, next) => {
   try {
-    let productById = await Product.findByPk(req.params.productId)
-    if (!productById) res.sendStatus(404)
-    else {
-      await Product.destroy({
-        where: {
-          id: req.params.productId
+    if (req.session.passport) {
+      const user = await User.findByPk(req.session.passport.user)
+      const isAnAdmin = user.isAdmin
+      if (isAnAdmin) {
+        let productById = await Product.findByPk(req.params.productId)
+        if (!productById) {
+          res.sendStatus(404)
+        } else {
+          productById = await Product.findByPk(req.params.productId)
+          if (!productById) res.sendStatus(404)
+          else {
+            await Product.destroy({
+              where: {
+                id: req.params.productId
+              }
+            })
+            res.sendStatus(204)
+          }
         }
-      })
-      res.sendStatus(204)
+      }
+    } else {
+      res.sendStatus(403)
     }
   } catch (error) {
     next(error)
@@ -58,9 +85,30 @@ router.delete('/:productId', async (req, res, next) => {
 
 router.post('/', async (req, res, next) => {
   try {
-    const newProduct = await Product.create(req.body)
-    //we should refactor later, anywhere we accept req.body - for security
-    res.status(201).send(newProduct)
+    if (req.session.passport) {
+      const user = await User.findByPk(req.session.passport.user)
+      const isAnAdmin = user.isAdmin
+      if (isAnAdmin) {
+        let productById = await Product.findByPk(req.params.productId)
+        if (!productById) {
+          res.sendStatus(404)
+        } else {
+          const formData = {}
+          if (req.body.name) {
+            formData.name = req.body.name
+          } else {
+            res.status(206).send('Name is required')
+          }
+          if (req.body.price) formData.price = req.body.price
+          if (req.body.description) formData.description = req.body.description
+          if (req.body.image) formData.image = req.body.image
+          const newProduct = await Product.create(formData)
+          res.status(201).send(newProduct)
+        }
+      }
+    } else {
+      res.sendStatus(403)
+    }
   } catch (error) {
     next(error)
   }
